requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.0

# HTTP & API
requests==2.31.0
urllib3==2.1.0

# Database
SQLAlchemy==2.0.23

# GeoIP
geoip2==4.7.11
maxminddb==2.5.0

# Configuration
python-dotenv==1.0.0

# Utilities
pytz==2023.3

